+++ src/Events/EventBus.ts
@@ -1,0 +1,16 @@
+import EventEmitter from 'eventemitter3';
+
+interface BoclipsPlayerEvents {
+  'boclips-player/playing': [
+    {
+      addonId: string;
+    }
+  ];
+}
+const eventEmitter = new EventEmitter<BoclipsPlayerEvents>();
+
+export class EventBus {
+  public static getEmitter = () => {
+    return eventEmitter;
+  };
+}